moves request out into its own module
